Added Match-3 functionality
Board.cs
---
Switched over to a 2D array to store tiles.
Simplified functions, added a lot of comments.
Replaces tiles that are deleted.
Updates each tiles neighbors when tiles are deleted.
---
Tile.cs
---
added checkNeighbors which checks for null items and removes them.

+++ Assets/Resources/Scripts/Board.cs
@@ -4,28 +4,33 @@ using System.Collections.Generic;
 
 
 public class Board : MonoBehaviour {
-	public List<Tile> tiles = new List <Tile>();
-	public int GridWidth;
-	public int GridHeight;
+
+	//Variables
+	private int GridWidth = 6;
+	private int GridHeight = 6;
 	public GameObject tilePrefab;
-	//public List<List<Tile>> tileBoard = new List<List<Tile>> ();
-	public List<Vector2> positions = new List<Vector2>();
+	public Tile[,] tileBoard = new Tile[6,6]; 
 	public List<Tile> tilesTouched = new List<Tile>();
+	public bool objectsRemoved;
 	bool trace = false;
 	bool bFirstTile = false;
 	string tileType = "";
 
-	// Use this for initialization
+
+	/// <summary>
+	/// void Start ()
+	/// 
+	/// Initializes the gameboard.
+	/// 
+	/// </summary>
 	void Start () {
 		for (int y = 0; y < GridHeight; y++) {
 			for (int x = 0; x < GridWidth; x++)
 			{
 				GameObject g = Instantiate(tilePrefab, new Vector3(x,y,0), Quaternion.identity)as GameObject;
 				g.transform.name = string.Format("Tile ({0},{1})",x,y);
-				//GameObject s = Instantiate(paperTilePrefab, new Vector3(x,y,0), new Quaternion(0,0,Random.Range(0,3)*90,0)) as GameObject;
 				g.transform.parent =gameObject.transform;
-				//s.transform.parent = gameObject.transform;
-				tiles.Add(g.GetComponent<Tile>());
+				tileBoard[x,y] = g.GetComponent<Tile>();
 			}
 				
 		}
@@ -33,23 +38,63 @@ public class Board : MonoBehaviour {
 		gameObject.transform.position = new Vector3 (-2.525f, -2.0f, 0);
 	}
 
+	/// <summary>
+	/// void Update()
+	/// 
+	/// track fingers and draw lines
+	/// 
+	/// </summary>
+	void Update()
+	{
+		handleInput ();
+		trackPosition ();
+		DrawLines();
+		tilesUpdateNeighbors ();
+		refillGameBoard ();
+	}
+
+
+	/// <summary>
+	/// 	void handleInput()
+	/// 
+	/// This function handles input from mouse/touch clicks and drags.
+	/// 
+	/// </summary>
+	void handleInput()
+	{
+		bool bMatched3 = (tilesTouched.Count >= 3);
+		//Starting click, start dragging
+		if (Input.GetMouseButtonDown (0) && !trace) {
+				trace = true;
+				bFirstTile = true;
+				//Ending click, end dragging, clear array, delete tiles
+		} else if (Input.GetMouseButtonUp (0) && trace && bMatched3) {
+				trace = false;
+				deleteAllTiles (tilesTouched);
+				tilesTouched.Clear ();
+				tileType = "";
+		} else if (Input.GetMouseButtonUp (0) && trace && !bMatched3) {
+				trace = false;
+				tilesTouched.Clear ();
+				tileType = "";
+		}
+
+	}
+	
+	/// <summary>
+	/// void trackPosition(void)
+	/// 
+	/// This function keeps tracks of which tiles are touched
+	/// 
+	/// </summary>
 	void trackPosition ()
 	{
 		int tileFoundIndex = 0;
 		Vector3 pos;
 		RaycastHit2D rayHit;
 		Tile tileFound;
-		if (Input.GetMouseButtonDown (0) && !trace) {
-			trace = true;
-			bFirstTile = true;
-		} else if (Input.GetMouseButtonUp (0) && trace) {
-			trace = false;
-			deleteAllTiles(tilesTouched);
-			tilesTouched.Clear();
-			tileType = "";
-			tileFoundIndex = 0;
-		}
 
+		//If we have started to trace
 		if (trace) {
 			bool bTileFound = false;
 			//Calculate position of mouse
@@ -60,25 +105,38 @@ public class Board : MonoBehaviour {
 			if (rayHit.collider != null) {
 				//set tileFound to rayHit Tile
 				tileFound = rayHit.transform.GetComponent<Tile> ();
+				//if this is the first tile
 				if(bFirstTile)
 				{
 					Debug.Log ("First Tile");
+					//Set the type of drag to this tile's type
 					tileType = tileFound.type;
+					//Add tile to the array
 					tilesTouched.Add (tileFound);
+					//indicate we are no longer dealing with the first tile
 					bFirstTile = false;
 				}
+				//if this is not the first tile
 				else{
-					tileFoundIndex = findTiles(tilesTouched, tileFound, ref bTileFound);
+					//search the touchedTiles for this tile
+					tileFoundIndex = findTiles(tilesTouched, tileFound);
+					//set boolean
+					if(tileFoundIndex >= 0)
+					{
+						bTileFound = true;
+					}
+					//if the tiletype matches
 					if(tileType == tileFound.type)
 					{
+						//if the tile is not found and is a neighbor of the last tile
 						if (!bTileFound && tileIsNeighbor(tileFound)) {
 						//add to array
 							Debug.Log ("Added Tile {0}", tileFound);
 							tilesTouched.Add (tileFound);
 						}
-						//else it is found in the array
+						//else it is found in the array, and it wasn't the last tile we dragged over
 						else if (bTileFound && tileIsNotLastTile(tileFound)) {
-						//find where in the array there is a tile
+							//remove all tiles up to the previous position of this tile in the array
 							Debug.Log(string.Format("Removing tiles {0} - {1}", tileFoundIndex, tilesTouched.Count - tileFoundIndex));
 							tilesTouched.RemoveRange (tileFoundIndex, tilesTouched.Count - tileFoundIndex);
 						}
@@ -90,10 +148,13 @@ public class Board : MonoBehaviour {
 				}
 			}
 		}
-		//}
-
 	}
 
+	/// <summary>
+	/// void DrawLines()
+	/// Simple debug function for drawing lines over the mouse/touch inputs.
+	/// 
+	/// </summary>
 	void DrawLines()
 	{
 		Vector3 pos1;
@@ -106,26 +167,41 @@ public class Board : MonoBehaviour {
 		}
 	}
 
-	int findTiles(List<Tile> tileArray, Tile toFind, ref bool bTileFound)
+	/// <summary>
+	/// int findTiles(List<Tile> tileArray, Tile toFind, ref bool bTileFound)
+	/// tileArray - any array of tiles
+	/// Tile toFind - the tile you are looking for in the array.
+	/// returns int - index of the array if found, otherwise -1
+	/// Simple debug function for drawing lines over the mouse/touch inputs.
+	/// 
+	/// </summary>
+	int findTiles(List<Tile> tileArray, Tile toFind)
 	{
-		int arrayIndex = 0;
-		foreach (Tile t in tileArray) {
-			//If the tile found is in the array
-			if (toFind.name == t.name) {
-				bTileFound = true;
-				break;
-			} else
-				arrayIndex++;
+		int arrayIndex = -1;
+		for (int i = 0; i < tileArray.Count; i++) {
+			if (toFind.name == tileArray [i].name) {
+					arrayIndex = i;
+					break;
+			}
 		}
 		return arrayIndex;
 	}
 
+	/// <summary>
+	/// bool tileIsNeighbor(Tile tileFound)
+	/// tileFound - tile to check
+	/// 
+	/// Function which checkts the last tile in the tilesTouched 
+	/// array to see if it is a neighbor
+	/// 
+	/// </summary>
 	bool tileIsNeighbor(Tile tileFound)
 	{
-		bool tileIsNeighbor = false;
 		Tile t = tilesTouched [tilesTouched.Count - 1];
-		findTiles (t.Neighbors, tileFound, ref tileIsNeighbor);
-		return tileIsNeighbor;
+		if (findTiles (t.Neighbors, tileFound) >= 0)
+				return true;
+		else
+				return false;
 	}
 
 	bool tileIsNotLastTile(Tile tileFound)
@@ -139,9 +215,23 @@ public class Board : MonoBehaviour {
 				Destroy (t.gameObject);
 			}
 	}
-	void Update()
+
+	void tilesUpdateNeighbors()
 	{
-		trackPosition ();
-		DrawLines();
+		for (int y = 0; y < GridHeight; y++)
+			for (int x = 0; x < GridWidth; x++)
+				tileBoard [x, y].checkNeighbors ();
+	}
+
+	void refillGameBoard()
+	{
+		for (int y = 0; y < GridHeight; y++)
+			for (int x = 0; x < GridWidth; x++)
+				if (tileBoard [x, y] == null) {
+				GameObject g = Instantiate (tilePrefab, new Vector3 (x-2.525f, y-2.0f, 0), Quaternion.identity)as GameObject;
+					g.transform.name = string.Format ("Tile ({0},{1})", x, y);
+					g.transform.parent = gameObject.transform;
+					tileBoard [x, y] = g.GetComponent<Tile> ();
+				}
 	}
 }
